backend/requirements.txt
Flask==2.3.3
pyserial==3.5
flask-cors==4.0.0
requests==2.31.0